@@ -1,6 +1,7 @@
 {
   "name": "youtube-transcript",
-  "version": "1.2.1",
+  "version": "1.3.0",
+  "type": "module",
   "description": "Fetch transcript from a youtube video",
   "main": "dist/youtube-transcript.common.js",
   "browser": "dist/youtube-transcript.esm.js",
@@ -16,11 +17,11 @@
   ],
   "license": "MIT",
   "devDependencies": {
-    "rollup": "^2.28.2",
+    "rollup": "^4.17.2",
     "rollup-plugin-typescript": "^1.0.1",
-    "rollup-plugin-typescript2": "^0.27.3",
-    "tslib": "^2.0.1",
-    "typescript": "^4.0.3"
+    "rollup-plugin-typescript2": "^0.36.0",
+    "tslib": "^2.6.2",
+    "typescript": "^5.4.5"
   },
   "files": [
     "dist/*"
@@ -32,5 +33,8 @@
   "homepage": "https://github.com/Kakulukian/youtube-transcript",
   "engines": {
     "node": ">=18.0.0"
+  },
+  "dependencies": {
+    "axios": "^1.6.8"
   }
 }
